Improvements
I made it so that we can reference a layer by its display name instead
of manually putting a number (I only applied this to KHS and KEA wind
turbines), and i also added a feature to make your own point colors for
each point layer (I changed some colors). All you have to do to change
the color is add the list of 3 rgb colors to the end of the Point_Layers
variable.